update db framework section
