+++ config/db.php.aux
@@ -7,5 +7,9 @@ return [
     'port' => '',
     'dbname' => '',
     'username' => '',
-    'password' => ''
+    'password' => '',
+    'options' => [
+        PDO::ATTR_ERRMODE => PDO::ERRMODE_EXCEPTION,
+        PDO::ATTR_DEFAULT_FETCH_MODE => PDO::FETCH_OBJ
+    ],
 ];
